*** empty log message ***
git-svn-id: https://wxmaxima.svn.sourceforge.net/svnroot/wxmaxima/trunk@391 cef4cce0-9ccc-409e-9180-6d2a30c8385f

+++ src/EditorCell.h
@@ -84,7 +84,7 @@ public:
   }
 private:
   wxString m_text;
-  long m_positionOfCaret;
+  unsigned int m_positionOfCaret;
   long m_selectionStart;
   long m_selectionEnd;
   int m_numberOfLines;
